move Help and Language from Footer to Header, moved style to _Header.css
